pyproject.toml
[project]
name = "tech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "pydantic-settings>=2.0.0",
    "nanoid>=2.0.0",
    "redis[hiredis]>=7.0.1",
    "tortoise-orm[asyncpg]>=0.25",
    "orjson>=3.11.4",
    "bcrypt>=5.0.0",
    "authx>=1.4.3",
    "authx-extra>=1.1.3",
]


[tool.ruff]
line-length = 128


[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
line-ending = "auto"
